tsapp/requirements.txt
gensim==3.8.3
nltk
networkx
numpy
pyrebase4